Make client better about incremental updates
Prior to this change, changing the Description field on, say, a CONSTANT
flag that had Enabled=true had the (very bad!) side effect of setting
Enabled=false. More broadly, the issue was that we were zeroing out
every field that was not explicitly set by the user on the command-line.

Now, we first do a lookup of the feature, and use
`(*pflag.FlagSet).Changed()` to detect which flags the user _actually_
set, and only update those.