Add Reader impl to undo forced html escaping in `json.Marshal`
In `encoding/json`, the encoder used by `json.Marshal` sets `escapeHTML:
true` [1], which means that any string literals containing angle
brackets or an ampersand (and the nonvisible line/paragraph separators)
will have the HTML-escaped equivalents rather than the literal
character, _even if_ the marshalled json is being used in non-html
contexts.

[1]: https://github.com/golang/go/blob/cca23a73733ff166722c69359f0bb45e12ccaa2b/src/encoding/json/encode.go#L161

It's _somewhat_ possible to work around this by using a custom encoder,
setting `encoder.SetEscapeHTML` and then calling `encoder.Encode`.
However, there are many situations where custom `json.Marshaler`
implementations get in the way of this, and encoding a nested structure
will still result in the nested structures being marshalled with escaped
HTML.

Therefore, this PR adds an HTMLUnescaper type that can be used as an
`io.Reader` that when reading bytes from it will undo the escaping of
the `json` package. The intent is to use this type with `io.Copy`, like:

    ```go
    data, err := json.Marshal(myvar)
    if err != nil {
        return nil
    }

    buf := bytes.NewBuffer(nil)
    _, err = io.Copy(buf, NewHTMLEscaper(data))
    ```